Trait novuskinc::console::KernelConsoleDriver

pub trait KernelConsoleDriver {
    fn write_character(&mut self, c: char, x: u16, y: u16) { ... }
    fn write_string(&mut self, string: &str, x: u16, y: u16) { ... }
    fn new_line(&mut self) { ... }
    fn clear_screen(&mut self, option: u16) { ... }
    fn dimensions(&mut self) -> (u16, u16) { ... }
}

The console driver has to implement KernelConsoleDriver because the functions the driver implements get called by the kernel

Provided Methods

fn write_character(&mut self, c: char, x: u16, y: u16)

The write_character function is used to write a single character to a certain place

fn write_string(&mut self, string: &str, x: u16, y: u16)

The write_string function is used to write a string and should call write_character to write the characters individually

fn new_line(&mut self)

The new_line function is used to move to the next line for the next character to be placed

fn clear_screen(&mut self, option: u16)

The clear_screen function is used to clear the kernel console, the option value is an optional argument, it can be used for whatever you want.

fn dimensions(&mut self) -> (u16, u16)

This should return the width and height field of KernelConsole.