Trait novuskinc::drivers::Driver

pub trait Driver: KernelConsoleDriver + FrameBufferGraphics + KeyboardInput + Storage {
    fn driver_name(&self) -> &'static str { ... }
    fn name(&self) -> &'static str { ... }
    fn init(&self) -> DriverResult { ... }
}

Provided Methods

fn driver_name(&self) -> &'static str

The driver_name function should be used to return the driver’s name

Example:

    fn driver_name(&self) -> &'static str {
        return "ANCD (A New Console Driver)";
    }

fn name(&self) -> &'static str

The name should return the “type” of driver it is so the kernel can easily index it in the driver manager

Example:

    fn name(&self) -> &'static str {
        return "Console Driver" // This is for a console driver
    }

This page has a list of driver names and what they’re used for.

fn init(&self) -> DriverResult

The init function needs to be used to initialize the driver, it is what the kernel calls to start it.